Test errors raised via misuse of `article_url`.
Various tests of `get article_url` and `patch article_url` that raise
ActionController::UrlGenerationError if the helper isn't called with
an argument, typically an Article or its id. One test is a test from
the Rails Testing Guide
(http://edgeguides.rubyonrails.org/testing.html)
which was referenced in the description of rails issue #31472.